Remove pre_exec for bash.
It's nice to have a timestamp upon issuing a command; but the problem is
that the output of the pre_exec function gets prepended to the output of
the first user-typed command. This is not a problem per se, but as soon
as we start playing around with pipes, this gets out of hand.

=> wait for a solution for the urxvt problem and change to zsh as soon
   as we make the terminal switch.