Add custom 6x13 font file.
The original 6x13 (Fixed SemiCondensed) seems to cause problemes with
dzen2. This is just an ungly hack to make it work again (long-term
solution is to dump dzen2 in favour for a homebrew solution, but until
then, months are likely to pass, and I don't want to wait that long for
a pretty statusbar).