Make a proper zsh config
The prompt is almost what it was like in bash, plus a few improvements;
notably the git indicator and the right prompt:

The git indicator turns red for untracked/unstaged files, otherwise
yellow for uncommitted files, otherwise cyan for unpushed commits,
otherwise blue.

The right prompt shows the time the last command took (if > 0s). Also,
it shows the return value for the last command (if != 0).

A few settings added by compinstall have been added, too.