Replace zshenv with zprofile; fix some file paths
zshenv is somewhat weird; it seems to be loaded at each invocation of a
shell, but is then again overwritten by /etc/profile (so it's probably
loaded *before* that). Although it's recommended not to export PATH and
other variables in zprofile, it's the only way it works (otherwise I'd
have to change files in /etc, and that really sucks).

.config/X11/xinitrc is generously ignored by startx, as it does not
honour XINITRC. xinit does not seem to work at all (which does seem to
have nothing to do with the location of .xinitrc, but much rather with
the "rootless X" problematic). The only way to let startx honour
.config/X11/xinitrc is to tell it up invocation.
... I'll probably make an alias for that crap.

Lastly, the dzenstat wrapper script does no longer source the Xresources
file automatically (just in order to keep things consistent; updating
the X resources database must happen on the user's request, also I won't
need to check ).

Oh, and I forgot to mention that I don't manually swap CapsLock and
Escape through xmodmap any longer, but let the -option 'caps:swapescape'
do the work. The xmodmap way somehow didn't work (need to track this
down).

Also, some minor changes, check the diff for more info.