Handle PATH on our own; add pacbloat scrpit
Long term solution is to source the scripts in /etc/profile.d manually,
thus more or less replacing most of /etc/profile. This allows to give
the user the full control over environment variables (especially PATH).

pacbloat simply lists all installed Arch Linux packages that have been
installed explicitely and that are not part of a dependency of another
package, or in either of the groups 'base' and 'base-devel'.

tmux now gets started automatically whenever there is an SSH session (so
no matter whether we are on a server or not). This simplifies a lot and
is also better.

urxvt only uses the default fixed-semicondensed (actually there was
never a need for the other font, as the special symbols we need are
included therein).