Fix PATH build-up, disable ZSH auto-menu
Go back to using the system-provided PATH, and extend it intelligently,
using a really ugly shell function that could have easily been replaced
by typeset -o path, then adding stuff to the $path array - but could not
because we keep compatibility with bash.

The zsh auto_menu feature is a little annoying because it completes
stuff that you don't want, among a choice of dozens of other options,
where the one you want is 99% a painful amount of TAB-ing away. It is
disabled.

Same goes for shared history, as my brain cannot support this.