Add bash_profile, additional checks for urxvtd, and adapt a few paths
bashrc is all that is needed in order to run bash, but at login it is
not sourced, which has not been noticed until I tried to log into a
session that was "still" using bash.

The bash history is now stored in $XDG_CACHE_HOME.

Upon starting X, we check for an already existing urxvtd before
murdering the current socket. This is due to urxvtcs not spawning after
a new X session had been launched, because the old socket had gone.

New muttrc location confuses the vim filetype plugin. [FIXED]