Fix escape delay in urxvt, add command line editing, add git merge state
Pressing [v] in the zsh command mode will open vim with the current
command line; save-quitting will put the content back to the command
line.

The delay caused by [esc] in zsh was caused by urxvt trying to interpret
the Escape key before passing it to the underlying process. We fix this
by passing on \033\007 (although I have no clue why exactly the bell
character should fix it, and why it worked fine for tmux and vim, but
not zsh).

[^r] in zsh insert mode will no longer start reverse search. I was shown
[/] in command mode ^^

In case one is in a merging state in a git repo, the prompt indicator
changes now accordingly.

The CDPATH has been removed. An example remains to show how it would
work if one intends to do so.