[vls] Handle shutdown gracefully
Given that the service file calls ExecStop whenever the process stops on
its own, we need to handle that case, otherwise it will translate to a
failed state, even if everything went correctly.

We let the service file call 'vls stop', which handles the case of no
process running gracefully. Hence the return to 'start' and 'stop'
commands in the script.