Add .zprofile to handle path variable under Arch Linux
The problem is that under Arch Linux, they have decided to source
/etc/profile (which sets the PATH variable) *after* having sourced
.zshenv. This has the funny effect of nuking your personal PATH variable
that you meticuously set in .zshenv (which, after all, should be the
place to define such things, right?).

Anyway, we do as in bash, we source the user specific shell environment
configuration from (a newly created) .zprofile now.

Fuck Arch.