Replace urxvtc by urxvt, add hint for karuibar implementation
If urxvtd crashes, *all* urxvtc instances become unusable. This is
pretty annoying (so annoying, in fact, that it was enough to piss me off
by happening only once).

Also, Xresources has an additional line that shows how me may handle
machine-dependent configuration for karuibar, together with the right
system call to obtain the host name, so that it should be able to
implement it in karuibar.