Add aacfy script that creates AAC versions of music files
iTunes is a major pain in the ass, as it somehow refuses to play certain
songs if they do not *exactly* match the format it wishes.

... heck, even the Finder preview is capable of playing stuff more
reliably than iTunes is.

As I often forget the syntax for using NeroAac, I put this in a script.
It's not like this won't happen ever again...