Attempt to tinker around in `checkrestart` script
The idea was to display the corresponding library files, too, but
since most processes may depend on multiple libraries, this is a project
suited for another time (not on an evening in a kitchen in the middle of
a stressful semester).

Also attempt to configure urxvt to use the "NotoMono" font for Emojis
(didn't work either).