Make checkrestart check for files in /usr
Checking for files with `lib` or `bin` has two disadvantages:
* It is pretty arbitrary.
* It may match stuff in your home.

Also, /usr is to be treated as read-only (when it comes to regular
user-space operations), so it's really just there where changes matter.

This update will not respect library files elsewhere than in /usr (e.g.
if the user has their own little library directory, e.g. in
~/.local/lib) - but the user is supposed to handle *those* on their own.