New approach to scripts
Utility scripts and programs are now in https://github.com/ayekat/utils.
To keep it tracked in a reasonable manner, we simply symlink that
repository's `bin` directory to ~/.local/bin/utils, and add it to the
$PATH.

Same applies to the dotfile's `bin` directory: it is now supposed to be
linked to ~/.local/bin/dotfiles.

Essentially, ~/.local/bin itself becomes untracked, which allows us to
install third-party stuff there without much hassle about what we want
to track and what not.