Add qutebrowser config
dwb kept crashing for a growing number of websites, so Firefox was often
used in parallel for all those sites.

But Firefox has increasingly become a major PITA, and I've reached a
point where my urge to get rid of Firefox became stronger than my desire
to remain with dwb.

So here we are. Let's see if we can get sane session support and
japanese input working with qutebrowser.