Wrapper scripts: Set path to the "real" executable correctly
The problem with setting $PATH directly is that this is propagated to
the exec'd process, too (TIL) - so if *that* process wants to run a
command, it does not take our dotfiles' PATH into account anymore.

> Example: claws-mail launched qutebrowser to open a link - given that
> claws-mail itself was running with a "stripped" PATH, it would not run
> the wrapper, but /usr/bin/qutebrowser directly.

This change sets EPATH instead, which denotes the full path to the
"next" executable. As far as I know, it should not interfere with any
functionality.