mimeapps.list: Add [Added Associations] section
For some reason, applications put themselves there, too, so I thought I
should just get ahead.

It's ugly, but in some way, it makes sense. It's just the format of the
file that is really ugly (and the fact that applications can modify that
file).

DON'T TOUCH MY BLOODY ETC!