Move .local sub-directories to top-level (PREPARE FOR BREAKAGE)
There is no point in keeping a single (hidden) directory that contains
the real content. Also, this allows using stuff in the repository
directly (e.g. for $PATH), as the structure is less fucked up.

Please note that probably all of your symbolic links are dead now.
Please fix them.