qutebrowser: Set LIBGL_ALWAYS_SOFTWARE on phobia
phobia is my stationary machine at home running with nouveau, which is
not supported by QtWebEngine.

Determining the host name requires the `hostname` command, which is part
of the `inetutils` package on Arch Linux, so the PKGBUILD has been
updated, too. It may be more elegant to replace host-specific
configuration by a detection of the GPU driver.