Remove XDG_TMP_HOME again
I think this is about the 3rd time I switch from one decision to the
other. For the decision this time: I think it's unnecessary complexity,
and even the system-wide /tmp (or /var/tmp) is a slight "special case"
(as in: stuff that is temporary but needs to be shared across processes
or preserved across sessions).

Let's see how long I'll stick with this ^^