Remove wrappers for ls, less and man
The problem with wrappers is that they remain even if the "real"
executable behind it does not exist. This may be worth opening an issue
(can't right now, sitting in a train).

Also, given their somewhat ugly nature, wrappers should only be used if
absolutely necessary, i.e. if they are not purely a command-line-only
application. Colour styling issues for ls and man don't fall into this
category.