ls: Incorporate `literal` quoting style into alias
We put this kind of configuration into sh/config and not sh/environment
(or even .pam_environment) because it's about pure console applications.
But for the same reason, it is perfectly fine to define an alias rather
than setting up an environment variable or using a wrapper script.

Also, environment variables polute the namespace.