etc/sh: Detect when we are in a bare git directory
It's still a bit ugly - we can't use `git status` inside a bare
repository, so we must find another way to track which remote branch
corresponds to the currently checked out one, and whether we are ahead
and/or behind.

Currently, the prompt just shows the branch name (indicating no changes,
even if there are) and no info about the relation to the remote branch
(indicating just "bare").