mutt: Rename to "mutt-hack"
Because - let's face it - it's a hack. And I'm currently working on a
saner mutt configuration (i.e. to be only used as MUA, in combination
with isync/mbsync and msmtp (thanks to @Phenitei for sharing his way of
doing it (still struggling with those pesky folder-hooks, though (and
exams, of course)))).