etc/mbsync: Propagate mailbox deletion both ways
A mailbox can be deleted by removing the `cur` directory. Note that this
propagation does not succeed if the remote mailbox still contains mails,
so in order to delete a Mailbox, there needs to be a first
synchronisation step between moving the mails out and deleting the
mailbox.

For "renaming" a directory, it's then basically:

1. Create mailbox directory containing `cur`, `new` and `tmp`
2. Move mails from the old mailbox here, preferrably with a MUA.
3. Synchronise
4. Delete `cur` in old mailbox directory
5. Synchronise
6. Delete mailbox directory

I'll have to see to what extent this can be scripted. The main issue is
scripting step 2 (but that should be doable - it's not like we don't
move mails around with pooch already ^^).