etc/pacman/pacfix: Add missing optdepends colons for `pass` rule
For visual pleasure, I initially intended to omit them and have pacfix
parse the file by interpreting the first word as the package name and
the remaining line as the description. But it turns out that this
`pkgname: some description`-format is used in the .PKGINFO as well, so I
can actually just dump the entire line as-is (and it's then consistent
with how it's done in PKGBUILDs, so probably easier for users).

Important trivia: `: ` (as in, colon followed by space) is the separator
between the package name and the reason. If there is only a colon or
only a space, pacman interprets the full line as the package name. Also,
having a line end with `: ` (as in, no description) is totally fine (in
such a case, not having `: ` is fine, too).