etc/newsboat: Remove feeds for packaging
The attempt was to tap into the bug tracker, but the issue is that the
feed only gives the 10 last updates each (and the 50 last for the
recently updated packages), so there is quite a high chance I might
occasionally miss some of the news, which kinda defeats the point.

Ikiwiki feed is also removed, since I don't really need to track
upstream, as I simply build the AUR package without modification
(although... it recently got orphaned - I hope someone capable picks it
up at some point).