etc/sh/profile: Fix criteria for adding directory to PATH
The previous criteria was "if it's a symlink, don't add it", assuming
that a symlink would exist solely for the purpose of the /usr and sbin
merge, which is a bit naive: if the target directory is not part of
PATH, it is never added.

Instead, we now attempt to add the target directory directly, and we
check for its existence as a directory.