pam_environment: Use ~/.local/data as XDG_DATA_HOME
XDG_DATA_HOME is now primarily intended for static/immutable data.
Ultimately, this is supposed to be in ~/.local/share (to align with the
FHS); however, there are still applications that don't properly
implement the XDG base directory specification, and this is a great way
to detect them. Notably, telegram-desktop hardcodes ~/.local/share
rather than using XDG_DATA_HOME, (see telegramdesktop/tdesktop#5741).
Until that is fixed, we'll keep XDG_DATA_HOME at this non-default
location.

For state data (analoguous to the FHS), we will use ~/.local/var/lib,
and introduce XDG_STATE_HOME, as discussed in issues #29 and #30.