etc/newsboat: Move cache file to $XDG_STATE_HOME
The name is slightly misleading. Essentially, the cache is the database
of locally stored article. While this does indeed qualify as a "cache",
the nature of RSS/Atom does not ensure that the remote will stay
available, and this "cache" cannot actually be regenerated (and thus
does not deserve the name).

Unfortunately, newsboat hardcodes the location of its history files, so
we can only move the database here, but not the commandline/search
history.