pam_environment: Set DBUS_SESSION_BUS_ADDRESS
It appears that it is sourced *after* pam_systemd exports both
variables, so DBUS_SESSION_BUS_ADDRESS is not set using XDG_RUNTIME_DIR
(at least not the one set by the user in .pam_environment), and we thus
need to set both explicitly.

This was never an issue on my machines, since our XDG_RUNTIME_DIR was
simply a symlink to /run/user/{uid}, which is the default for
XDG_RUNTIME_DIR anyway. But that is no reason not to be explicit about
it (especially for the (unlikely) case where we decide to effectively
use a different, non-default directory as our XDG_RUNTIME_DIR).

Thanks to Tom^ on freenode.net#archlinux for pointing this out!