bin: Remove darktable wrapper script
It's not exactly the first application to misunderstand XDG_CONFIG_HOME,
and we've already got a simple and lightweight solution for that:
gitignore.

This makes wrapper scripts worse than applications dumping non-config
files into XDG_CONFIG_HOME, so I'll opt for the lesser evil here (as a
reminder: wrapper scripts pollute the namespace and tab completion if
the backed program is not installed).

Oh, and it turns out XDG_CONFIG_HOME/darktable was actually already in
the gitignore.