Have dotfiles reside in ~/.local directly [BREAKING CHANGE]
This is mainly to solve issue #12 (so we no longer have to care about
where the repo resides, because we just put *everything* into the repo),
but it also solves the symlink hell we've gotten ourselves into: the
only symbolic links required now are pam_environment, and the symlink
for XDG_RUNTIME_DIR.

On a sidenote, this is probably about as megalomaniac as this repository
could become. We could go a step further and put all of $HOME into this
repo, but since we've got all our "dotfiles" in ~/.local already, what's
the point anyway?