etc/sh: Move files from config to profile.d
In the end, this file is not exactly useful. Environment variables are
propagated to non-interactive shells if run after logging in, and we
don't expect non-interactive shells to run in a not-logged-in
environment (if that happens, we have to treat it specially anyway, so
might as well assume it never happens).