Remove mnt
I find it weird to have a (single) pre-existing location for mounting
drives; I think it's better to just create them ad-hoc for whatever
usecase (which I already do in most cases). Ultimately, it feels like
keeping around the mnt directory would just be for having ~/.local
resemble the FHS more, which is a bit silly.