etc/X11/xinitrc.d/10-xinput.sh: Reorganise + fix for libinput 1.13.3
For some reason, the mouse acceleration is super low, so I need to ramp
it up manually. Also, rewrite the snippet to use xinput to remap the
mouse buttons (so drop xmodmap as a requirement).

Also make the device->id mapping a bit more elegant (instead of just a
blunt and rather unreliable grep).