Add arandr wrapper
arandr hardcodes the location of its layout directory. The annoying part
is that it auto-creates ~/.screenlayout each time I want to open a saved
layout, and drops me in there, so that I have to manually navigate to
the correct location for loading the layout (also, have dotbloat delete
~/.screenlayout afterwards).

The workaround now is the extremely ugly "fake home hack": we pretend
that our home directory is $XDG_CONFIG_HOME/arandr, and we let it do
what it wants to do in there (requires me to rename the `layouts`
directory to `.screenlayout` directory in there, but I can live with
that).

I mean, it worked fine for firefox as well...