etc/zsh: Don't print user name in prompt at all
There is no point in defining a "default user", as the localdir
configuration is not effective as another user anyway (if properly
switched to another user, that is). So ultimately, it's just a matter of
maintaining a list of "On that machine, I am user XYZ" in this
configuration, which is a bit silly.