etc/zr: Remove debootstrap from [repkg]
The main issue (bashisms in #!/bin/sh script) has been fixed a while
ago, and the minor issue (not using pacman-conf) has been fixed just
now.

There's still missing dependencies for this package, but those are sh
and pacman, and it's not worth maintaining a repkg'd version of
debootstrap for that.