Salt: Make it user-only properly
There is this variable that is not mentioned anywhere in the example
master configuration file, sqlite_queue_dir. This led to the master
still trying to create /var/cache/salt/master/queues, making it fail.

Also, turn lib/salt/bin/salt into a more generic script, and symlink the
other salt commands to it.