etc/vim: Don't use viminfo for Git commit messages
I haven't enabled "jump back to previous position", but some distros
and/or sysadmins like to do so system-wide.

I don't care too much about it, but it's rather annoying for Git commit
messages.

Also adapt disabling the undo file to be more in line with where we
disable viminfo (i.e. all of /dev and all of /run, no just the …/shm
variants).