etc/sh: Force ALSA on gurke to use internal speakers by default
gurke is often attached to a ThinkPad Thunderbolt dock, which provides
its own audio card. Unfortunately, this causes ALSA to randonly switch
to that as a default card, which many of my applications can't handle
(since I'm not using PulseAudio yet), and forces me to control audio on
the (non-default) card through alsamixer.

Until I've migrated to PulseAudio, this is a hacky workaround to prevent
that.