wrappers: Remove sudo
It's quite a bit of an "oof" security-wise. I've got into the habit of
running `sudoedit`, and I'm avoiding situations where I actually have to
edit files via pacdiff by repkg'ing packages in a way that I can use
dropin snippets (pacman is still a todo), so... this should no longer be
needed.