etc/X11/Xresources.d/urxvt: Revert font setting change
Turns out the issue was caused by urxvt not finding the font because
there was no fonts.dir index file in /usr/share/fonts/misc. This entry
is usually created by mkfontdir (part of the `xorg-mkfontscale` package
in Arch), which was not installed.

We now repkg rxvt-unicode to declare that missing dependency;
subsequently, we can revert to loading the desired font in urxvt.