MPD: Use Pulseaudio via TCP - harder than it should be
It's actually quite simple:

* Pulseaudio has the ability to listen for audio streams on a TCP socket
  (with module-native-protocol-tcp), on port 4713.
* MPD has the ability to stream audio to Pulseaudio via TCP.

So essentially, we only need to write $XDG_CONFIG_HOME/pulse/default.pa
that loads the desired module, and configure mpd to stream to the PA TCP
socket, right?

Unfortunately, Pulseaudio does not respect XDG_CONFIG_HOME for reading
configuration files (it does so only for the runtime and state files,
but ironically *not* the configuration); it simply searches for a
hardcoded ~/.config/pulse/ there.

I believe this is a bug, but who actually fucking knows these days...

So! As a workaround, I could create a dummy ~/.config, with a symlink to
my default.pa, but that's ugly. As a less ugly workaround (quite
fortunately, actually), I can explicitly tell Pulseaudio which
default.pa to use, via PULSE_SCRIPT (hence the additional change in
~/.pam_environment).

And last (and rather least), since we previously ignored the entirety of
$XDG_CONFIG_HOME/pulse, we need to adapt the .gitignore file(s) there a
bit, too.

*phew*

In any case, now I can finally use MPD for multi-user setups :-)