etc/gtk-{2.0,3.0}: Dark theme
For gtk-2.0, use Adwaita-dark. For gtk-3.0, stick to the light theme for
now (since the dark theme is a bit painful there), but at least add the
line to remind me how to select it.

For gtk-3.0, also remove the gtk-color-scheme option, as it is
deprecated and ignored now. Unfortunately my Git history doesn't tell me
why I added it.