Move variable declaration from pam_environment to environment.d
Because environment.d only applies to the systemd user instance (but not
the user session), we additionally need to process those files in the
login shell now. In the long term, we may want to move more variables to
environment.d (not just the ones for XDG).

We still keep pam_environment around for declaring XDG_CONFIG_HOME (to
solve the chicken-or-egg issue for systemd-environment-d-generator) and
ZDOTDIR (to solve the chicken-or-egg issue for zsh).

In cases where pam_env.so no longer loads ~/.pam_environment (either
because of local configuration, or because of pam_env.so dropping
support for that entirely), the work-around is to symlink ~/.config to
~/.local/lib/dotfiles/config (which then contains a single symlink for
environment.d) and symlink ~/.zlogin to ~/.local/etc/zsh/.zlogin.

An alternative would be to kindly ask the system administrator to set
XDG_CONFIG_HOME and ZDOTDIR to desired values for my user (though I
still have to find the most elegant way to achieve that).