wrappers/unwrap: Handle case where the executed file is a symlink
The goal here is to avoid issues if we have wrapper scripts that are
symlinks of other wrapper scripts.

The problem was that which() gave back the path to the *resolved* file,
which isn't great in cases like `ansible-playbook`, or any other
applications that care about under which name they've been invoked.