etc/zr: Add zuepfe-updates repo
This is for keeping around updates to packages that are in the official
repositories. Not quite the same as zuepfe-repkg, because our goal isn't
to change things, but merely to get up-to-date software.

This will cause some updates in pacman-hacks (diffrepo) as well, and
potentially also in the repkg'd pacman package.