Cleanup after remove gurke
The company laptop is now running a system that is completely detached
from the zuepfe ecosystem. "Spiritually" speaking, gurke is gone.

This also means that we can drop quite a significant amount of
application configuration and packages.

See also the removals in [zuepfe-original] in the PKGBUILDs repository
on GitLab and in the private localdir and gitolite-admin repositories on
git.zuepfe.net.