share: Invert ignoring
It's time we take back some control over this directory. The first step
is to assume we want to track what is inside here, and treat the
directories we don't track as exceptions (similar to etc and lib).

Essentially, applications should drop their stuff into one of the
subdirectories of var, not here.