Update irssi wrapper script to be more versatile
Overall overhaul of the script (shebang, shell options, using readonly
variables, ...).

Furthermore, don't expect there to be one specific password. Not
everyone connects to the zuepfe.net bouncer, and not everyone is ayekat.

To continue working, the irssi configuration file(s) must be adapted to
replace

    ZNC_PASSWORD

by

    @IRC_PASSWORD:some/password@

where `some/password` is the password location (under `services`) in the
password store.