lib/karuiwm/scrlock: Add case for slock
While we would like to keep private and work environments separate,
that's mostly for the OS layer (to avoid Zuepfe depending on a
technically untrusted admin device). For localdir, I think it's fine to
have *some* work elements bleed into the private environment for now.

More broadly, the cleanest way would probably be a `karuiwm-scrlock`
virtual package provided by two different (mutually conflicting)
packages `karuiwm-scrlock-xtrlock` and `karuiwm-scrlock-slock`, and the
environment-specific meta-packages (or even just the user at install
time) deciding which one to pull in. But that's something to be fixed in
Zuepfe, so for now we'll have to stick to this workaround.