etc/X11/Xresources: Fix DPI to 96
When attaching external monitors, some newly opened windows use a higher
DPI value (though I'm not quite sure where they take that information
from), which results in the cursor being oversized in those windows.

It's also very inconsistent because it only applies to newly opened
windows (so e.g. not the root window or anything that's been opened
before).

Here, we hardcode the X11 server DPI settings to the "normal"
(not-hiDPI) 96 DPI.