Add talky-ssh-agent, a wrapper around ssh-agent that sends notifications
The idea is to be notified of anything that interacts with an SSH agent.

This is a sort of compromise solution for not having to type the SSH key
passphrase each time, but also not just running an SSH agent without any
tracking of what's happening.

This doesn't add security in the sense of preventing things from
accessing the SSH key, but at least it gives me a notification whenever
something shady is going on.