Git: Try to fix gpg silliness when invoked via SSH/TTY
Initially I was about to blame GPG and/or Pinentry, but it turns out Git
is invoking Pinentry with not TTY attached, so obviously gpg-agent
cannot prompt for a passphrase in ncurses mode.

Also, Git doesn't treat the argument to gpg.command as a pathname, so
tilde expansion doesn't work there (and subsequently I need to place
this wrapper script in $PATH). I would have preferred this in
$XDG_LIB_HOME/git, but welp.

Blame Git. Twice.