etc/karuiwm: Re-add old ThinkPad keyboard for further testing
Had a theory that it might not be the buttons, but the cable.

So now I'm using it with a different cable. Let's try to reduce waste.

(though the issue is that just yesterday, I briefly had the same button
issue with the new cable---I'll try a little further, but the next time
this drives me insane, this keyboard is gone for real).