Remove xdg-open wrapper, declare XDG_CURRENT_DESKTOP globally
I guess my initial thought here was that this affects only xdg-mime, so
it should be put in a dedicated wrapper rather than polluting the entire
environment.

But really all these environment variables are already polluting my
environment anyway, as they're merely hacks/workarounds for applications
not properly adhering to the XDG basedir spec.

I need to find out if this really needs to be GNOME, though (upon
cursory inspection of xdg-mime, I could not find anything that suggests
it needs to be GNOME and cannot be anything else/more fitting like e.g.
`generic`).