sh/profile.d/10-environment-d: Run generators instead of manual parsing
As this assumes there are generators, this also adds systemd as a
dependency. But in 2022, I think it's fine to assume that's now pretty
much universal.

And if I happen to land on a system where systemd (or at least the
environment generators) are not present, I'll revert to manual parsing.