etc/vim: Show window-local modified state in inactive window
If we use &modified as a reference, we only get the focused window's
modified state, which leads to the wrong indicator being shown.

The double space is because... honestly, I have no idea. A single space
had no effect, so I randomly tried with a second one, and now it works.

If anyone is able to explain this to me, please please please do so.