vim workaround
Visual formatting is broken since the upgrade from 9.1.0000 to 9.1.0080
(can't tell where exactly this happened): Background gets a dark grey
and foreground gets just black. Text is practically illegible this way.

This is a workaround fix (no formatting from my side ends up with the
default, which is still fine). The only difference is that the text has
no syntax highlighting if part of the selection, but the background is
the same otherwise.