Mark qutebrowser wrapper script as unmaintained
With the previous commit (switching to qbsmgr), we no longer need this
script for managing our sessions.

Note that this script will be removed in the following commit; this
commit here merely allows me to link to this script for reference, but
to also be clear that it's no longer maintained.